Array search
Create a function that receives an array of integers as parameter, and one integer value that has to be found in the array! Return the index of the element if the value was found, or return −1, if it was not. (If the value occurs multiple times in the array, it is up to you which index you return.)