Array separation
Write a function that receives an integer array, and separates the even and the odd elements of the array. The even elements are put into a second array, and the odd ones are put into a third array; both of them are function arguments, and are assumed to be large enough to store all even and odd elements.